Change API of the benchmark rpc client.
Right now it will use channels, as it would allow to spawn your own
tasks outside which would bring more flexibility.
