+++ crates/subspace-farmer/src/bin/subspace-farmer/bench_rpc_client.rs
@@ -1,14 +1,8 @@
-use std::sync::atomic::{AtomicU64, Ordering};
 use std::sync::Arc;
-use std::time::Duration;
 
 use async_trait::async_trait;
-use rand::prelude::*;
 use subspace_archiving::archiver::ArchivedSegment;
-use subspace_core_primitives::objects::{PieceObject, PieceObjectMapping};
-use subspace_core_primitives::{
-    ArchivedBlockProgress, BlockNumber, FlatPieces, LastArchivedBlock, RootBlock, Sha256Hash,
-};
+use subspace_core_primitives::BlockNumber;
 use subspace_rpc_primitives::{
     BlockSignature, BlockSigningInfo, FarmerMetadata, SlotInfo, SolutionResponse,
 };
@@ -32,100 +26,43 @@ pub struct Inner {
     archived_segments_receiver: Arc<Mutex<mpsc::Receiver<ArchivedSegment>>>,
     slot_info_handler: Mutex<JoinHandle<()>>,
     segment_producer_handle: Mutex<JoinHandle<()>>,
-    writen_piece_count: Arc<AtomicU64>,
 }
 
 impl BenchRpcClient {
     /// Create a new instance of [`BenchRpcClient`].
-    pub fn new(metadata: FarmerMetadata) -> Self {
+    pub fn new(
+        metadata: FarmerMetadata,
+    ) -> (Self, mpsc::Sender<ArchivedSegment>, mpsc::Sender<SlotInfo>) {
         let (slot_info_sender, slot_info_receiver) = mpsc::channel(10);
         let (archived_segments_sender, archived_segments_receiver) = mpsc::channel(10);
         let (acknowledge_archived_segment_sender, mut acknowledge_archived_segment_receiver) =
             mpsc::channel(1);
-        let writen_piece_count = Arc::new(AtomicU64::new(0));
+        let (outer_archived_segments_sender, mut outer_archived_segments_receiver) =
+            mpsc::channel(10);
+        let (outer_slot_info_sender, mut outer_slot_info_receiver) = mpsc::channel(10);
 
         let slot_info_handler = tokio::spawn(async move {
-            let mut slot_number = 0;
-            let mut next_salt = rand::random();
-            loop {
-                let global_challenge = rand::random();
-                next_salt = {
-                    let (salt, next_salt) = (next_salt, rand::random());
-                    let slot_info = SlotInfo {
-                        slot_number,
-                        global_challenge,
-                        salt: next_salt,
-                        next_salt: salt,
-                        solution_range: rand::random(),
-                    };
-                    if slot_info_sender.send(slot_info).await.is_err() {
-                        break;
-                    };
-                    Some(next_salt)
+            while let Some(slot_info) = outer_slot_info_receiver.recv().await {
+                if slot_info_sender.send(slot_info).await.is_err() {
+                    break;
                 };
-
-                tokio::time::sleep(Duration::from_secs(2 * 60)).await;
-
-                slot_number += 1;
             }
         });
 
         let segment_producer_handle = tokio::spawn({
-            let writen_piece_count = Arc::clone(&writen_piece_count);
             async move {
-                let mut segment_index = 0;
-                let mut last_archived_block = LastArchivedBlock {
-                    number: 0,
-                    archived_progress: ArchivedBlockProgress::Partial(0),
-                };
-                loop {
-                    last_archived_block
-                        .archived_progress
-                        .set_partial(segment_index as u32);
-
-                    let archived_segment = {
-                        let root_block = RootBlock::V0 {
-                            segment_index,
-                            records_root: Sha256Hash::default(),
-                            prev_root_block_hash: Sha256Hash::default(),
-                            last_archived_block,
-                        };
-
-                        let mut pieces = FlatPieces::new(100);
-                        rand::thread_rng().fill(pieces.as_mut());
-
-                        let objects = std::iter::repeat_with(|| PieceObject::V0 {
-                            hash: rand::random(),
-                            offset: rand::random(),
-                        })
-                        .take(100)
-                        .collect();
-
-                        ArchivedSegment {
-                            root_block,
-                            pieces,
-                            object_mapping: vec![PieceObjectMapping { objects }],
-                        }
-                    };
-
-                    if archived_segments_sender
-                        .send(archived_segment)
-                        .await
-                        .is_err()
-                    {
+                while let Some(segment) = outer_archived_segments_receiver.recv().await {
+                    if archived_segments_sender.send(segment).await.is_err() {
                         break;
                     }
                     if acknowledge_archived_segment_receiver.recv().await.is_none() {
                         break;
                     }
-
-                    segment_index += 1;
-                    writen_piece_count.fetch_add(100, Ordering::AcqRel);
                 }
             }
         });
 
-        Self {
+        let me = Self {
             inner: Arc::new(Inner {
                 metadata,
                 slot_info_receiver: Arc::new(Mutex::new(slot_info_receiver)),
@@ -133,13 +70,9 @@ impl BenchRpcClient {
                 acknowledge_archived_segment_sender,
                 slot_info_handler: Mutex::new(slot_info_handler),
                 segment_producer_handle: Mutex::new(segment_producer_handle),
-                writen_piece_count,
             }),
-        }
-    }
-
-    pub fn writen_piece_count(&self) -> u64 {
-        self.inner.writen_piece_count.load(Ordering::Relaxed)
+        };
+        (me, outer_archived_segments_sender, outer_slot_info_sender)
     }
 
     pub async fn stop(self) {
